[PATCH] msm: acpuclock-8960: Initialize regulators and bus_bw earlier
The init code assumes that the regulators have already been set
up to support the max CPU frequency. While this is true, the
fewer assumptions we make, the better. So, switch the ordering to
initialize the regulators before switching the frequency. This
also has the nice side-effect of not running the regulators at a
higher than required voltage for fast parts.

While we are at it, also change the bus init ordering so that we
can pick the bus bandwidth that corresponds to the freq the L2
will be initialized to instead of picking the max bandwidth
defined in the bandwidth table.

Change-Id: I58432f6e236a7077536e6565157b96c309df590d
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>