[PATCH] msm: acpuclock-8960: Confirm power-collapse before disabling CPU clocks
Hotplug may not always be configured to involve power-collapse of
the CPU. When it is not, the CPU must remain clocked. If an HFPLL
is still selected as the clock source of the removed CPU, do not
disable it as part of the CPU_DEAD hotplug notification unless
msm_pm_verify_cpu_pc() confirms that the CPU has been collapsed.

This code can be significantly simplified if support for Krait v1
processors is ever removed from the driver. Krait v2 and later
allow changing the CPU clock source from another CPU even after
the affected CPU is collapsed. This feature could be used to change
the collapsed CPU's source away from the HFPLLs to something that
is always on (ex. QSB).

Change-Id: Ifee795a336f1cf9b0cc5c883c2bea23cb1043ccd
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>