[PATCH] msm: acpuclock-8960: Skip regulator on first hfpll disable
When the HFPLLs are first initialized we don't want to direct the
RPM to disable the regulators for the HFPLLs until all the PLLs
are configured. Otherwise we'll end up in a situation where an
HFPLL is enabled but the regulator for the HFPLL is disabled.

For example, when the first HFPLL is initialized we may be in a
state where both CPUs are running at 918MHz using their
respective HFPLLs. When hfpll_init() disables the first CPUs PLL
to reconfigure its speed it will disable the PLL with
hfpll_disable(). The regulator driver will see that there are no
other votes for the regulator (even though the other CPU is using
its own HFPLL) and then disable the supply used for all the
HFPLLs.

Skip disabling the regulator on the first hfpll_disable() so that
we properly initialize all the votes for the regulators from the
PLLs and only drop the votes once all votes are initialized. If
the regulator APIs were in use there would have been an
unbalanced disable warning printed.

Change-Id: Id8c13e3076845d39b46a903296ed61d96490521d
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>