[PATCH] msm: acpuclock-8960: Fix regulator-related debug prints
Regulator names are not printed in all regulator-related debug messages.
Fix this and remove the duplicate __func__ prefix for these messages.

Also, the regulator names 'krait0_dig' and 'krait0_mem' were assigned for
all CPUs, not just Krait 0.  Change this so the number in the regulator
name matches the CPU number the regulator is associated with.

Change-Id: I1843d30670c44434f8f190a2b8d7fda71c308357
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>