[PATCH] msm: board-8960: Change PM8921 SMPS 8 constraint voltages to allow 2.05 V
Hardware characterization has shown that the system operates
without any problem when the voltage of SMPS 8 is lowered from
2.1 V to 2.05 V.  Therefore, change the constraint range for
SMPS 8 to include 2.05 V as a lower bound.

At the same time, change the voltage requests of the two SMPS 8
consumers: acpuclock-8960 and clock-8960 to use the new, lower
voltage value.

Change-Id: Ic3084acf9a69e81ace0f71c10d0cbdc35b81112f
Signed-off-by: David Collins <collinsd@codeaurora.org>